Add tile sync and sample tile
