+++ datalayer/sample/wear/src/debug/java/com/google/android/horologist/datalayer/sample/SampleTilePreview.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.horologist.datalayer.sample
+
+import android.content.Context
+import androidx.wear.tiles.tooling.preview.Preview
+import androidx.wear.tiles.tooling.preview.TilePreviewData
+import androidx.wear.tooling.preview.devices.WearDevices
+import com.google.android.horologist.compose.tools.tileRendererPreviewData
+
+@Preview(device = WearDevices.LARGE_ROUND)
+@Preview(device = WearDevices.SMALL_ROUND)
+fun SampleTilePreview(context: Context): TilePreviewData = tileRendererPreviewData(
+    renderer = SampleTileRenderer(context),
+    tileState = Unit,
+    resourceState = Unit,
+)
